requirements.txt
bs4==0.0.1
decorator==4.4.2
email-validator==1.1.3
google-api-core==2.8.0
google-api-python-client==2.48.0
google-auth==1.28.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1
googleapis-common-protos==1.56.1
oauth2client==4.1.2
oauthlib==3.1.0
rich==12.4.4
typer==0.3.2
typing-extensions==4.2.0
virtualenv==20.0.17

